requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
cryptography
azure-functions
azure-storage-blob
azure-identity
requests
pandas
numpy